+++ src/utils/paths.py
@@ -1,15 +1,22 @@
 from contextlib import contextmanager
 from pathlib import Path
+import pathlib
 import os
 import sys
 from rich.console import Console
 console=Console()
+
+from ..constants import configPath
+from ..utils import profiles
+
+homeDir=pathlib.Path.home()
+
 @contextmanager
 def set_directory(path: Path):
     """Sets the cwd within the context
 
-    Args:
-        path (Path): The path to the cwd
+        Args:``
+            path (``Path): The path to the cwd
 
     Yields:
         None
@@ -29,3 +36,25 @@ def createDir(path):
     except:
         console.print("Error creating directory, check the directory and make sure correct permissions have been issued.")
         sys.exit()
+def databasePathHelper(model_id,username):
+    profile = profiles.get_current_profile()
+    return homeDir / configPath / profile / ".data"/f"{username}_{model_id}"/"user_data.db"
+
+
+def messageResponsePathHelper(model_id,username):
+    profile = profiles.get_current_profile()
+    return homeDir / configPath / profile / ".data"/f"{username}_{model_id}"/"messages.json"
+
+
+def timelineResponsePathHelper(model_id,username):
+    profile = profiles.get_current_profile()
+    return homeDir / configPath / profile / ".data"/f"{username}_{model_id}"/"timeline.json"
+
+
+def archiveResponsePathHelper(model_id,username):
+    profile = profiles.get_current_profile()
+    return homeDir / configPath / profile / ".data"/f"{username}_{model_id}"/"archive.json"
+def pinnedResponsePathHelper(model_id,username):
+    profile = profiles.get_current_profile()
+    return homeDir / configPath / profile / ".data"/f"{username}_{model_id}"/"pinned.json"
+